feat: refactor RemoteController and enhance JSON usage - Import JSON and update imports related to typing and file paths in `main.py`. - Introduce usage of `anyio` library in `main.py`. - Add win/lose attributes and related methods to the `RemoteController` class. - Remove `GameResult` class inheritance from `RemoteController` and delete the `game.py` file. - Implement new computed property `win_lost_dict` to calculate win and lose statistics in `RemoteController`. - Create an asynchronous method to export win/lose statistics to a JSON file in `RemoteController`.
